--- a/mmdet/core/bbox/coder/yolo_bbox_coder.py
+++ b/mmdet/core/bbox/coder/yolo_bbox_coder.py
@@ -70,21 +70,13 @@ class YOLOBBoxCoder(BaseBBoxCoder):
         Returns:
             torch.Tensor: Decoded boxes.
         """
-        assert pred_bboxes.size(0) == bboxes.size(0)
         assert pred_bboxes.size(-1) == bboxes.size(-1) == 4
-        x_center = (bboxes[..., 0] + bboxes[..., 2]) * 0.5
-        y_center = (bboxes[..., 1] + bboxes[..., 3]) * 0.5
-        w = bboxes[..., 2] - bboxes[..., 0]
-        h = bboxes[..., 3] - bboxes[..., 1]
-        # Get outputs x, y
-        x_center_pred = (pred_bboxes[..., 0] - 0.5) * stride + x_center
-        y_center_pred = (pred_bboxes[..., 1] - 0.5) * stride + y_center
-        w_pred = torch.exp(pred_bboxes[..., 2]) * w
-        h_pred = torch.exp(pred_bboxes[..., 3]) * h
-
+        xys = (bboxes[..., :2] +
+               bboxes[..., 2:]) * 0.5 + (pred_bboxes[..., :2] - 0.5) * stride
+        whs = (bboxes[..., 2:] -
+               bboxes[..., :2]) * 0.5 * pred_bboxes[..., 2:].exp()
         decoded_bboxes = torch.stack(
-            (x_center_pred - w_pred / 2, y_center_pred - h_pred / 2,
-             x_center_pred + w_pred / 2, y_center_pred + h_pred / 2),
+            (xys[..., 0] - whs[..., 0], xys[..., 1] - whs[..., 1],
+             xys[..., 0] + whs[..., 0], xys[..., 1] + whs[..., 1]),
             dim=-1)
-
         return decoded_bboxes