--- a/mmdet/core/bbox/coder/yolo_bbox_coder.py
+++ b/mmdet/core/bbox/coder/yolo_bbox_coder.py
@@ -71,12 +71,13 @@ class YOLOBBoxCoder(BaseBBoxCoder):
             torch.Tensor: Decoded boxes.
         """
         assert pred_bboxes.size(-1) == bboxes.size(-1) == 4
-        xys = (bboxes[..., :2] +
-               bboxes[..., 2:]) * 0.5 + (pred_bboxes[..., :2] - 0.5) * stride
+        xy_centers = (bboxes[..., :2] + bboxes[..., 2:]) * 0.5 + (
+            pred_bboxes[..., :2] - 0.5) * stride
         whs = (bboxes[..., 2:] -
                bboxes[..., :2]) * 0.5 * pred_bboxes[..., 2:].exp()
         decoded_bboxes = torch.stack(
-            (xys[..., 0] - whs[..., 0], xys[..., 1] - whs[..., 1],
-             xys[..., 0] + whs[..., 0], xys[..., 1] + whs[..., 1]),
+            (xy_centers[..., 0] - whs[..., 0], xy_centers[..., 1] -
+             whs[..., 1], xy_centers[..., 0] + whs[..., 0],
+             xy_centers[..., 1] + whs[..., 1]),
             dim=-1)
         return decoded_bboxes