## Motivation

Speed up yolov3 inference by refactoring the `get_bboxes` logic.

## Modification

To speed up the inference time, I refactored the following part:
1. Move onnx export logic to `onnx_export` function. Simplify the `get_bboxes` logic. Speed up yolov3 mbv2 320 about 3 FPS.
2. Improve code parallelism. Replace the for loop of multi-level feature map decode with a flatten tensor decode. Speed up yolov3 mbv2 320 about 6 FPS.
3. Simplify the bbox decode in yolo bbox coder. Speed up yolov3 mbv2 320 about 4 FPS.

Finally improved about 10FPS.

* *Note: Tested on a single 1060 6G GPU with torch1.8 cuda11.1*

## Results

|           | yolov3 d53 608 | yolov3 d53 416 | yolov3 d53 320 | yolov3 mbv2 416 | yolov3 mbv2 320 |
|---------|----------------|----------------|----------------|-----------------|-----------------|
| Before  | 23.2fps        | 44.9fps        | 53.2fps        | 79.7fps         | 93.9fps         |
| After     | 24.2fps        | 48.5fps        | 57.4fps        | 91.5fps         | 106.2fps        |

## BC-breaking (Optional)

None